tighten review trust boundaries and stage dispatch
